feat: Implement ranked analysis with budgeting and retries
This commit introduces the `trigger-ranked-analysis` CLI command, which automates procurement analysis based on a specified budget and priority ranking.

Key Features:
- **Ranked Analysis CLI (`trigger-ranked-analysis`):** A new command to trigger analyses based on a budget. It supports a `--zero-vote-budget-percent` to limit spending on non-voted items and a `--max-messages` parameter to limit the number of triggered analyses.
- **Priority Ranking:** Analyses are prioritized based on vote count, recency, warnings, and cost.
- **Budget Ledger:** A new `budget_ledgers` table tracks all expenses related to analysis jobs for financial transparency.
- **Resilient API Calls:** A `retry_with_backoff` decorator has been implemented and applied to all external API calls (PNCP and Gemini) to handle transient network errors gracefully.
- **Comprehensive Testing:** The E2E test has been updated to validate the new `trigger-ranked-analysis` flow, and all unit tests have been fixed to reflect the new logic. The entire test suite now passes reliably.
- **Documentation:** A detailed technical design document (`docs/intelligent_budget_analysis.md`) explains the new features and logic.

Database Changes:
- The `votes_count` column has been moved from `procurements` to `procurement_analyses`.
- The precision of the `amount` column in `donations` and `budget_ledgers` has been increased to `DECIMAL(18, 10)` to handle precise cost calculations.

+++ source/repositories/budget_ledger.py
@@ -3,6 +3,7 @@ This module defines the repository for handling database operations
 related to the budget ledger.
 """
 
+from decimal import Decimal
 from uuid import UUID
 
 from models.budget_ledger import TransactionType
@@ -21,7 +22,7 @@ class BudgetLedgerRepository:
         self.logger = LoggingProvider().get_logger()
         self.engine = engine
 
-    def save_expense(self, analysis_id: UUID, cost: float, description: str) -> None:
+    def save_expense(self, analysis_id: UUID, cost: Decimal, description: str) -> None:
         """Saves a new expense record to the budget ledger.
 
         Args:
